+++ source/whalecar-service/src/main/java/com/whalecar/domain/WxCarPrice.java
@@ -1,18 +1,20 @@
 package com.whalecar.domain;
 
+import java.math.BigDecimal;
+
 /**
  * Created by wufei on 14/11/23.
  */
 public class WxCarPrice {
 
 
-    public Integer getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 
-    private Integer price;
+    private BigDecimal price;
 }
